+++ spec/models/renalware/events/event_type_alert_trigger_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+module Renalware::Events
+  describe EventTypeAlertTrigger, type: :model do
+    it :aggregate_failures do
+      is_expected.to validate_presence_of(:event_type)
+    end
+  end
+end